requirements.txt
# The officially supported versions of the following packages are their
# python3-* equivalent distributed in Ubuntu 22.04 and Debian 11
Babel==2.9.1  # min version = 2.6.0 (Focal with security backports)
chardet==4.0.0
cryptography==3.4.8
decorator==4.4.2
docutils==0.17
ebaysdk==2.1.5
freezegun==1.1.0
geoip2==2.9.0
gevent==21.8.0 ; python_version == '3.10'  # (Jammy)
gevent==22.10.2; python_version > '3.10'
greenlet==1.1.2 ; python_version == '3.10'  # (Jammy)
greenlet==2.0.2 ; python_version > '3.10'
idna==2.10  # requests 2.25.1 depends on idna<3 and >=2.5
Jinja2==3.0.3 ; python_version <= '3.10'
Jinja2==3.1.2 ; python_version > '3.10'
libsass==0.20.1
lxml==4.8.0 ; python_version <= '3.10'
lxml==4.9.2 ; python_version > '3.10'
MarkupSafe==2.0.1 ; python_version <= '3.10'
MarkupSafe==2.1.2 ; python_version > '3.10'
num2words==0.5.10
ofxparse==0.21
passlib==1.7.4 # min version = 1.7.2 (Focal with security backports)
Pillow==9.0.1 ; python_version <= '3.10'  # min version = 7.0.0 (Focal with security backports)
Pillow==9.4.0 ; python_version > '3.10'
polib==1.1.1
psutil==5.9.0 ; python_version <= '3.10' 
psutil==5.9.4 ; python_version > '3.10' 
psycopg2==2.9.2 ; sys_platform != 'win32' and python_version <= '3.10'
psycopg2==2.9.5 ; python_version > '3.10' or sys_platform == 'win32'
pydot==1.4.2
pyopenssl==21.0.0
PyPDF2==1.26.0 ; python_version <= '3.10'
PyPDF2==2.12.1 ; python_version > '3.10'
pypiwin32 ; sys_platform == 'win32'
pyserial==3.5
python-dateutil==2.8.1
python-ldap==3.4.0 ; sys_platform != 'win32'  # min version = 3.2.0 (Focal with security backports)
python-stdnum==1.17
pytz  # no version pinning to avoid OS perturbations
pyusb==1.2.1
qrcode==7.3.1
reportlab==3.6.8 ; python_version <= '3.10'
reportlab==3.6.12 ; python_version > '3.10'
requests==2.25.1 # versions < 2.25 aren't compatible w/ urllib3 1.26. Bullseye = 2.25.1. min version = 2.22.0 (Focal)
rjsmin==1.1.0
urllib3==1.26.5 # indirect / min version = 1.25.8 (Focal with security backports)
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.*
zeep==4.1.0
pylint_odoo==9.0.0
pandas==2.1.2
openpyxl==3.1.3
